backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]
SQLAlchemy==2.0.35
python-dotenv==1.0.1
authlib==1.3.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.12.0
dspy-ai==3.0.1
itsdangerous==2.2.0


